+++ package.json
@@ -8,7 +8,7 @@
     "typechain": "typechain --target ethers-v5 --out-dir ./typechain 'external-artifacts/*.json'",
     "format": "prettier --check \"./**/*.{js,ts}\"",
     "format:fix": "prettier --write \"./**/*.{js,ts}\"",
-    "lint": "eslint --ext .ts test",
+    "lint": "eslint --ext .ts test deployment",
     "lint:fix": "yarn lint --fix",
     "typecheck": "tsc --noEmit --incremental false --composite false",
     "clean": "hardhat clean",
