+++ tsconfig.json
@@ -7,6 +7,6 @@
     "outDir": "dist",
     "declaration": true
   },
-  "include": ["./scripts", "./test", "./typechain"],
+  "include": ["./scripts", "./test", "./typechain", "./deployment"],
   "files": ["./hardhat.config.ts"]
 }
